refactor(ui): create api service
