+++ client/home.html
@@ -46,7 +46,18 @@
       {
         let s = document.createElement('script');
         s.src = "home.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+        s.onreadystatechange = s.onload = () => { window.homeInit(); };
+        ons.getScriptPage().onShow = function() {
+            if (window.homeInit) {
+                window.homeInit();
+            }
+        };
+
+        ons.getScriptPage().onHide = function() {
+            window.homeHide();
+        };
       }
     </script>
     <style>
