+++ client/index.js
@@ -48,100 +48,6 @@ window.fn.getTopPage =
             return document.getElementById("appNavigator").topPage;
         };
 
-window.fn.request = function(url, method, callback) {
-    var request = new XMLHttpRequest();
-    request.open(method, url, true);
-
-    request.onload = function() {
-        if (request.status >= 200 && request.status < 400) {
-            try {
-                callback(null, JSON.parse(request.responseText));
-            } catch (err) {
-                callback(null, request.responseText);
-            }
-        } else {
-            console.error(request);
-            callback("There was an error: " + request.status);
-        }
-    };
-
-    request.onerror = function() {
-        callback("Connection error");
-    };
-
-    request.send();
-};
-
-window.fn.requestWithPayload = function(url, payload, method, callback) {
-    var request = new XMLHttpRequest();
-    request.open(method, url, true);
-    request.setRequestHeader("Content-Type", "application/json; charset=UTF-8");
-    request.send(payload);
-
-    request.onload = function() {
-        if (request.status >= 200 && request.status < 400) {
-            try {
-                callback(null, JSON.parse(request.responseText));
-            } catch (err) {
-                callback(null, request.responseText);
-            }
-        } else {
-            console.error(request);
-            callback("There was an error: " + request.status);
-        }
-    };
-
-    request.onerror = function() {
-        callback("Connection error");
-    };
-};
-
-window.fn.postFile =
-    function(url, path, progressCallback, callback) {
-        var formData = new FormData();
-        formData.append("file", path);
-
-        var request = new XMLHttpRequest();
-        request.onerror = function(e) {
-            console.error(request);
-            callback("There was an error: " + request.status);
-        };
-
-        request.onload = function(e) {
-            if (request.status >= 200 && request.status < 400) {
-                try {
-                    callback(null, JSON.parse(request.responseText));
-                } catch (err) {
-                    callback(null, request.responseText);
-                }
-            } else {
-                console.error(request);
-                callback("There was an error: " + request.status);
-            }
-        };
-
-        request.upload.onprogress = function(e) {
-            var p = Math.round(100 / e.total * e.loaded);
-            progressCallback(p);
-        };
-
-        request.onerror = function() {
-            callback("Connection error");
-        };
-
-        request.open("POST", url);
-        request.send(formData);
-    };
-
-window.fn.secondsToHms = function(d) {
-    d = Number(d);
-
-    var h = Math.floor(d / 3600);
-    var m = Math.floor(d % 3600 / 60);
-    var s = Math.floor(d % 3600 % 60);
-
-    return ("0" + h).slice(-2) + ":" + ("0" + m).slice(-2) + ":" + ("0" + s).slice(-2);
-};
 window.fn.applyTheme =
     function() {
         let themeEl = document.getElementById("theme");
