+++ client/manualcontrol.html
@@ -13,7 +13,16 @@
       {
         let s = document.createElement('script');
         s.src = "manualcontrol.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            ManualControlInit();
+        };
+
+        ons.getScriptPage().onHide = function() {
+            ManualControlHide();
+        };
       }
     </script>
     <style>
