+++ client/settings-access-control.html
@@ -96,7 +96,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-access-control.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsAccessControlPage();
+        };
       }
     </script>
 </ons-page>
