+++ client/settings-carpet-mode.html
@@ -63,7 +63,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-carpet-mode.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsCarpetModePage();
+        };
       }
     </script>
 </ons-page>
