+++ client/settings-cleaning-history.html
@@ -18,7 +18,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-cleaning-history.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsCleaningHistoryPage();
+        };
       }
     </script>
 </ons-page>
