+++ client/settings-consumables.html
@@ -89,7 +89,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-consumables.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsConsumablesPage();
+        };
       }
     </script>
     <style>
