+++ client/settings-info.html
@@ -114,7 +114,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-info.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsInfoPage();
+        };
       }
     </script>
 </ons-page>
