+++ client/settings-mqtt.html
@@ -133,7 +133,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-mqtt.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsMqttPage();
+        };
       }
     </script>
