+++ client/settings-persistent-data.html
@@ -100,7 +100,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-persistent-data.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsPersistentDataPage();
+        };
       }
     </script>
 </ons-page>
