+++ client/settings-sound-voice.html
@@ -61,7 +61,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-sound-voice.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            InitSettingsSoundVolumePage();
+        };
       }
     </script>
 </ons-page>
