+++ client/settings-token.html
@@ -36,7 +36,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-token.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsTokenPage();
+        };
       }
     </script>
 </ons-page>
