+++ client/settings-wifi.html
@@ -75,7 +75,12 @@
       {
         let s = document.createElement('script');
         s.src = "settings-wifi.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            updateSettingsWifiPage();
+        };
       }
     </script>
