+++ client/zones.html
@@ -23,7 +23,12 @@
       {
         let s = document.createElement('script');
         s.src = "zones.js";
+        s.type = "module";
         ons.getScriptPage().appendChild(s);
+
+        ons.getScriptPage().onShow = function() {
+            ZonesInit();
+        };
       }
     </script>
     <style>
