+++ jsconfig.json
@@ -9,5 +9,5 @@
         "resolveJsonModule": true
     },
     "exclude": ["node_modules", "**/node_modules/*"],
-    "include": ["lib/**/*.js", "*.js", "lib/*.js"]
+    "include": ["lib/**/*.js", "*.js", "lib/*.js", "client/services/*.js"]
 }
